+++ noodles-bam/src/bai/async/reader.rs
@@ -1,3 +1,5 @@
+use std::collections::HashMap;
+
 use noodles_bgzf as bgzf;
 use noodles_csi::{
     index::{
@@ -142,7 +144,7 @@ where
     Ok(ReferenceSequence::new(bins, intervals, metadata))
 }
 
-async fn read_bins<R>(reader: &mut R) -> io::Result<(Vec<Bin>, Option<Metadata>)>
+async fn read_bins<R>(reader: &mut R) -> io::Result<(HashMap<usize, Bin>, Option<Metadata>)>
 where
     R: AsyncRead + Unpin,
 {
@@ -154,7 +156,7 @@ where
         usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
     })?;
 
-    let mut bins = Vec::with_capacity(n_bin);
+    let mut bins = HashMap::with_capacity(n_bin);
     let mut metadata = None;
 
     for _ in 0..n_bin {
@@ -166,8 +168,9 @@ where
             metadata = read_metadata(reader).await.map(Some)?;
         } else {
             let chunks = read_chunks(reader).await?;
-            let bin = Bin::new(id, bgzf::VirtualPosition::default(), chunks);
-            bins.push(bin);
+            let bin = Bin::new(bgzf::VirtualPosition::default(), chunks);
+            // TODO: Check for duplicates.
+            bins.insert(id, bin);
         }
     }