+++ noodles-csi/src/reader.rs
@@ -1,4 +1,5 @@
 use std::{
+    collections::HashMap,
     io::{self, Read},
     str,
 };
@@ -231,7 +232,7 @@ where
     Ok(reference_sequences)
 }
 
-fn read_bins<R>(reader: &mut R, depth: u8) -> io::Result<(Vec<Bin>, Option<Metadata>)>
+fn read_bins<R>(reader: &mut R, depth: u8) -> io::Result<(HashMap<usize, Bin>, Option<Metadata>)>
 where
     R: Read,
 {
@@ -239,7 +240,7 @@ where
         usize::try_from(n).map_err(|e| io::Error::new(io::ErrorKind::InvalidData, e))
     })?;
 
-    let mut bins = Vec::with_capacity(n_bin);
+    let mut bins = HashMap::with_capacity(n_bin);
 
     let metadata_id = Bin::metadata_id(depth);
     let mut metadata = None;
@@ -257,8 +258,9 @@ where
             metadata = read_metadata(reader).map(Some)?;
         } else {
             let chunks = read_chunks(reader)?;
-            let bin = Bin::new(id, loffset, chunks);
-            bins.push(bin);
+            let bin = Bin::new(loffset, chunks);
+            // TODO: Check for duplicates.
+            bins.insert(id, bin);
         }
     }